change return type to float

+++ src/distance.nim
@@ -1,4 +1,4 @@
 import propositionalLogic
 
-proc drasticDistance*(x, y: PropLogicFormula): int =
-  if x == y: 0 else: 1
+proc drasticDistance*(x, y: PropLogicFormula): float =
+  if x == y: 0.0 else: 1.0
